pyproject.toml
[project]
name = "s3-tables-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.74",
    "pyiceberg[duckdb,sql-sqlite]>=0.8.1",
    "python-dotenv>=1.0.1",
    "s3fs>=0.4.2",
]
